requirements.txt
# RAP Estimate Generator - Python Dependencies

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# PDF Processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
tabula-py==2.9.0
camelot-py[cv]==0.11.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# OCR (if needed for scanned PDFs)
pytesseract==0.3.10

# OpenAI (optional - for intelligent parsing fallback)
openai==1.6.1

# HTTP requests
requests==2.31.0

# Validation
pydantic==2.5.3

# File handling
python-multipart==0.0.6

# Logging
colorlog==6.8.0